better events on supervisor tree
